CLI: Keep the worker-thread TypeScript loader
Reverts the in-thread module.registerHooks loader: on Node versions
before the fix that landed between 22.22.1 and 22.22.3, CJS modules
imported through in-thread hooks receive a require without .extensions,
which crashes tsconfig-paths (Object.keys(require.extensions)) in the
react-docgen Vite plugin during every sandbox build. The worker-thread
module.register loader does not intercept in-process CJS and is immune.

Reproduced with the react-vite sandbox on Node 22.21.0: in-thread hooks
crash all docgen story transforms, the worker loader builds cleanly.
In-thread hooks can return behind a runtime probe once the fixed Node
versions are pinned down; they were worth ~50-150ms per invocation.

Also switches the csf-utils test mock to Vitest spy mode per review.
