+++ code/core/src/shared/utils/module.ts
@@ -1,17 +1,11 @@
-import { readFileSync, statSync } from 'node:fs';
-import NodeModule, { createRequire, register } from 'node:module';
+import { statSync } from 'node:fs';
+import { createRequire, register } from 'node:module';
 import { win32 } from 'node:path/win32';
 import { fileURLToPath, pathToFileURL } from 'node:url';
 
 import { resolveModulePath } from 'exsolve';
 import { dirname, join } from 'pathe';
 
-import {
-  addExtensionsToRelativeImports,
-  clearDirectoryCache,
-  isTypeScriptUrl,
-} from '../../bin/loader-utils.ts';
-import { NODE_TARGET } from '../constants/environments-support.ts';
 import { jsModuleExtensions } from '../constants/extensions.ts';
 
 /**
@@ -54,82 +48,6 @@ export const resolvePackageDir = (
 let isTypescriptLoaderRegistered = false;
 
 /**
- * Register the TypeScript loader. On Node with in-thread hook support (`module.registerHooks`,
- * 22.15+) and native type stripping (`module.stripTypeScriptTypes`, 22.13+) the transform runs
- * synchronously in-process. The `module.register` worker-thread loader is strictly more expensive:
- * every module load in the process — plain JS included — pays a synchronous IPC round-trip to the
- * hook worker, which profiles as 100–200ms per CLI invocation. Older Node falls back to it.
- */
-function registerTypescriptLoader() {
-  const { registerHooks, stripTypeScriptTypes } = NodeModule as unknown as {
-    registerHooks?: (hooks: {
-      load: (
-        url: string,
-        context: unknown,
-        nextLoad: (url: string, context?: unknown) => unknown
-      ) => unknown;
-    }) => void;
-    stripTypeScriptTypes?: (
-      code: string,
-      options?: { mode?: 'strip' | 'transform'; sourceMap?: boolean; sourceUrl?: string }
-    ) => string;
-  };
-
-  if (typeof registerHooks === 'function' && typeof stripTypeScriptTypes === 'function') {
-    // The native transform is not a full esbuild replacement: it only downlevels enums and
-    // namespaces, and future Node versions may reduce it to erase-only. Anything it rejects is
-    // retried with the same esbuild transform the worker loader applies, so the two registration
-    // paths accept the same input.
-    const transformTypeScript = (source: string, filePath: string): string => {
-      try {
-        // The source map keeps stack-trace locations of transformed config files accurate.
-        return stripTypeScriptTypes(source, {
-          mode: 'transform',
-          sourceMap: true,
-          sourceUrl: pathToFileURL(filePath).href,
-        });
-      } catch {
-        const { transformSync } = createRequire(import.meta.url)(
-          'esbuild'
-        ) as typeof import('esbuild');
-        return transformSync(source, {
-          loader: 'ts',
-          target: NODE_TARGET,
-          format: 'esm',
-          platform: 'neutral',
-        }).code;
-      }
-    };
-
-    registerHooks({
-      load(url, context, nextLoad) {
-        // Strip any query string (e.g. the cache-busting `?<timestamp>` importModule appends for
-        // skipCache) before checking the extension.
-        const urlWithoutQuery = url.split('?')[0];
-        if (!isTypeScriptUrl(urlWithoutQuery)) {
-          return nextLoad(url, context);
-        }
-        const filePath = fileURLToPath(urlWithoutQuery);
-        const rawSource = readFileSync(filePath, 'utf-8');
-        return {
-          format: 'module',
-          shortCircuit: true,
-          // Add extensions to relative imports so Node.js ESM can resolve them
-          source: addExtensionsToRelativeImports(
-            transformTypeScript(rawSource, filePath),
-            filePath
-          ),
-        };
-      },
-    });
-    return;
-  }
-
-  const typescriptLoaderUrl = importMetaResolve('storybook/internal/bin/loader');
-  register(typescriptLoaderUrl, import.meta.url);
-}
-
-/**
  * Dynamically imports a module with TypeScript support, falling back to require if necessary.
  *
  * @example Import a TypeScript preset
@@ -151,14 +69,9 @@ export async function importModule(
   { skipCache = false }: { skipCache?: boolean } = {}
 ) {
   if (!isTypescriptLoaderRegistered) {
-    registerTypescriptLoader();
+    const typescriptLoaderUrl = importMetaResolve('storybook/internal/bin/loader');
+    register(typescriptLoaderUrl, import.meta.url);
     isTypescriptLoaderRegistered = true;
-  }
-
-  if (skipCache) {
-    // A cache-busted re-import must also see current directory contents, or the loader's
-    // extension resolution can miss files created since the previous import.
-    clearDirectoryCache();
   }
 
   let mod;
